osxkeyboardcontrol: Shift event handling out of event tap callback
This should prevent us from ever seeing the "disabled due to timeout"
warning.

Key events are enqueued as `("key", True)` for a key up.
A method running in a loop in a worker thread dequeues these
and routes them to the appropriate `key_up` or `key_down` callback.

Teardown happens by setting a `_canceled` property and then dispatching
a `None` event to wake up the thread in case it's blocked on an empty
queue.